[fix] kakao unlink post 추가.
탈퇴 시 kakao과 연결을 끊도록 하는 unlink 코드를 추가했습니다.

+++ src/services/UserService.ts
@@ -1,12 +1,30 @@
-
 import User from '../models/User';
-
-
+import axios from 'axios';
+import errorGenerator from '../errors/errorGenerator';
+import statusCode from '../modules/statusCode';
 
 
 const deleteUser = async (userId: string) => {
     try {
+
+        const user = await User.findById(userId);
+
+        if (!user) {
+            throw errorGenerator({
+                msg: '유저 정보를 불러올 수 없습니다.',
+                statusCode: statusCode.NOT_FOUND
+            })
+        }
+
+        const result = await axios.post("https://kapi.kakao.com/v2/user/unlink", { 
+            headers: {
+                Authorization: `Bearer ${user.accessToken}`,
+            },
+        });
+
         await User.findByIdAndDelete(userId);
+
+        return result
     } catch (error) {
         console.log(error);
         throw (error);
